gmail-cleaner-hello/requirements.txt
# Gmail API and Google Authentication
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-auth==2.23.4

# Email handling and formats
email-validator==2.1.0

# Data storage and serialization
PyYAML==6.0.1
cryptography==41.0.7

# CLI and progress handling
click==8.1.7
tqdm==4.66.1
rich==13.6.0

# Database
sqlalchemy==2.0.23

# Date and time handling
python-dateutil==2.8.2

# File handling and compression
pathlib-extensions==0.1.0

# Testing (for future development)
pytest==7.4.3
pytest-mock==3.12.0

# Development tools
black==23.10.1
flake8==6.1.0
